Some fixes
Fixed glitch where flares on both sides will disappear when one side goes beyond the borders of the window; next task should be to adjust aspect ratio before adding more complicated features.

Bug still exists of flares of enemies disappearing as soon as they are killed, but too lazy to fix that and it will barely impact anyone or anything.